expense-ai-backend/requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
diff-match-patch==20241021
dj-database-url==3.1.2
Django==6.0.3
django-apscheduler==0.7.0
django-cors-headers==4.9.0
django-import-export==4.4.0
et_xmlfile==2.0.0
git-filter-repo==2.47.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
openpyxl==3.1.5
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
